# apps/cloud-backend/fusion-core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  types.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
    type: object
...
  github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProjectArchiveRequest:
    properties:
      is_archived:
        example: true
        type: boolean
    required:
    - is_archived
    type: object
  github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProjectCreateRequest:
    properties:
      application:
        example: Audio System Design
        maxLength: 255
        minLength: 1
        type: string
      budget:
        $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.Budget'
      description:
        example: Professional audio system for corporate conference room
        maxLength: 1000
        type: string
      environment_type:
        allOf:
        - $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.EnvironmentType'
        example: indoor
      is_project_file_created:
        example: false
        type: boolean
      is_project_thumbnail_created:
        example: false
        type: boolean
      name:
        example: Conference Room Audio Setup
        maxLength: 255
        minLength: 1
        type: string
      project_id:
        example: 50000001-0000-4000-8000-000000000008
        type: string
      project_phase:
        allOf:
        - $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProjectPhase'
        example: Proposal
      venue:
        example: Building A - Conference Room 101
        maxLength: 255
        type: string
    required:
    - application
    - environment_type
    - name
    - project_id
    type: object
  github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProjectCreateResponse:
    properties:
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      project_upload_url:
        example: https://storage.example.com/upload/projects/123e4567
        type: string
      thumbnail_upload_url:
        example: https://storage.example.com/upload/projects/123e4567
        type: string
    type: object
...
  github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProjectUpdateRequest:
    properties:
      application:
        example: Audio System Design
        maxLength: 255
        minLength: 1
        type: string
      budget:
        $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.Budget'
      description:
        example: Updated professional audio system for corporate conference room
        maxLength: 1000
        type: string
      environment_type:
        allOf:
        - $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.EnvironmentType'
        example: indoor
      is_project_file_dirty:
        example: true
        type: boolean
      is_project_thumbnail_dirty:
        example: true
        type: boolean
      name:
        example: Updated Conference Room Audio Setup
        maxLength: 255
        minLength: 1
        type: string
      project_phase:
        allOf:
        - $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProjectPhase'
        example: Development
      venue:
        example: Building B - Conference Room 205
        maxLength: 255
        type: string
    type: object
  github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProjectUpdateResponse:
    properties:
      project_upload_url:
        example: https://storage.example.com/upload/projects/123e4567
        type: string
      thumbnail_upload_url:
        example: https://storage.example.com/upload/projects/123e4567
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Fusion Cloud Backend API server providing comprehensive
    role-based access control, user management, and project management capabilities.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Fusion Cloud Backend API
  version: "1.0"
paths:
...
  /projects:
    get:
      consumes:
      - application/json
      description: Get all projects in the system
      parameters:
      - description: Filter projects by archived status
        in: query
        name: is_archived
        type: boolean
      - description: Field to sort projects by (e.g., created_at, updated_at)
        in: query
        name: sort_by
        type: string
      - description: Sort order (ascending or descending)
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved all projects
          schema:
            $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.GetAllProjectsResponse'
        "400":
          description: Bad request - Invalid query parameters
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Create a new project in the system
      parameters:
      - description: Project details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProjectCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created project
          schema:
            $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProjectCreateResponse'
        "400":
          description: Bad request - Invalid payload or user not found or project
            Id already exists
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new project
      tags:
      - projects
  /projects/{projectId}:
    delete:
      consumes:
      - application/json
      description: Delete a project by its unique identifier
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successfully deleted project
        "400":
          description: Bad request - Missing user ID
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden - User not assigned to project, project archived,
            or locked by another user
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Project or user not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - projects
    patch:
      consumes:
      - application/json
      description: Update an existing project by its ID
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Updated project details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProjectUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated project
          schema:
            $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProjectUpdateResponse'
        "400":
          description: Bad request - Invalid payload
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden - User not assigned to project, project archived,
            or locked by another user
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Project or user not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - projects
  /projects/{projectId}/archive:
    post:
      consumes:
      - application/json
      description: Archive or unarchive a project based on request body
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Archive/unarchive request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProjectArchiveRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Project archive status updated successfully
        "400":
          description: Bad request - Invalid payload
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden - User not assigned to project or project locked
            by another user
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update project archive status
      tags:
      - projects
  /projects/{projectId}/lock:
    post:
      consumes:
      - application/json
      description: Lock or unlock a project for a specific user based on request body
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Lock/unlock request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProjectLockRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Successfully updated project lock status
        "400":
          description: Bad request - Invalid payload
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden - User not assigned to project or project already
            locked by another user
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update project lock status
      tags:
      - projects
  /projects/{projectId}/star/{userId}:
    post:
      consumes:
      - application/json
      description: Star or unstar a project for a specific user based on request body
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Star/unstar request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.ProjectStarRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Successfully updated project star status
        "400":
          description: Bad request - Invalid payload
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: Forbidden - User not assigned to project, project archived,
            or locked by another user
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Project or User not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update project star status
      tags:
      - projects
  /projects/{projectId}/users/{userEmail}:
    delete:
      consumes:
      - application/json
      description: Remove a user from a project by project ID and user email
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: User Email
        in: path
        name: userEmail
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: User successfully removed from the project
        "404":
          description: Project or User not found, or user not assigned to the project
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove a user from a project
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Assign a user to a project by project ID and user email
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: User Email
        in: path
        name: userEmail
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: User successfully assigned to the project
        "404":
          description: Project or User not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assign a user to a project
      tags:
      - projects
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"